requirements/common.txt
# requirements/common.txt
Django==1.5.1
MySQL-python==1.2.4
beautifulsoup4==4.1.3
dj-database-url==0.2.1
django-debug-toolbar==0.9.4
django-tastypie==0.9.14
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2